chore: refactor kube ingress route columns
Refactors the Kubernetes ingress route columns using /columns and a props.ts.

Part of #11035.
