+++ packages/renderer/src/lib/ingresses-routes/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import IngressRouteActions from '../IngressRouteActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<IngressRouteActions ingressRoute={object} />
