+++ packages/renderer/src/lib/ingresses-routes/columns/HostPath.svelte
@@ -1,13 +1,9 @@
 <script lang="ts">
 import { Link } from '@podman-desktop/ui-svelte';
 
-import { IngressRouteUtils } from './ingress-route-utils';
-import type { IngressUI } from './IngressUI';
-import type { RouteUI } from './RouteUI';
+import { IngressRouteUtils } from '../ingress-route-utils';
+import type { Props } from './props';
 
-interface Props {
-  object: IngressUI | RouteUI;
-}
 let { object }: Props = $props();
 
 const ingressRouteUtils = new IngressRouteUtils();
